### 2018-07-24 #843 @whedon
[PRE REVIEW]: The Glasgow Fortran Source-to-Source Compiler
**Submitting author:** @wimvanderbauwhede (<a href="http://orcid.org/0000-0001-6768-0037">Wim Vanderbauwhede</a>)
**Repository:** <a href="https://github.com/wimvanderbauwhede/RefactorF4Acc" target ="_blank">https://github.com/wimvanderbauwhede/RefactorF4Acc</a>
**Version:** v1.0.0
