### 2018-07-25 c0a300c
Added paper.md
