### 2018-07-25 #843 @wimvanderbauwhede
Sorry, the paper was in my devel branch, I've now added it to the master branch.
