### 2018-07-31 #865 @wimvanderbauwhede
OK, I have checked the proof and it looks fine to me.

Wim
