### 2018-08-07 #865 @labarba
👋 @rouson, @rasmussn — Thank you for agreeing to review this JOSS submission. You can find a description of the [Review Criteria](https://joss.readthedocs.io/en/latest/review_criteria.html), and more, in our documentation pages. Feel free to ask any questions here. For specific change requests, you can open issues on the submission's repository, and mention in this thread, for a link-back.
