### 2018-08-21 d9009f5
Added modules to manipulate CPP macros
