+++ README.md
@@ -45,16 +45,16 @@ School of Computing Science, University of Glasgow, UK
 
 * Subroutine extraction  
     * simply add an annotation
-    
+
           !$ACC SUBROUTINE <optional subroutine name>  
           ... <code to be extracted as subroutine>
           !$ACC END SUBROUTINE <optional subroutine name>
-                   
+
 * Call graph generation
 
 ### Automatic parallelisation using OpenCL
 
-Automatic parallelisation and offloading of legacy code to accelerators is the ultimate aim of the project, and already works for many cases. 
+Automatic parallelisation and offloading of legacy code to accelerators is the ultimate aim of the project, and already works for many cases.
 However, the work flow is more complicated and requires an additional compiler, [AutoParallel-Fortran](https://github.com/wimvanderbauwhede/AutoParallel-Fortran). This compiler is written in Haskell, which is not yet a mainstream programming language. Furthermore, the generated OpenCL code relies on the [OclWrapper Fortran OpenCL API](https://github.com/wimvanderbauwhede/OpenCLIntegration), written in C++, and uses [scons](http://scons.org/), a Python-based build system. For these reasons, it is harder to install this autoparallelising compiler. However, if you have installed it, a test case for the full flow is provided in the `tests` folder, see below for more details.           
 
 ## How it works
@@ -100,7 +100,7 @@ To install RefactorF4ACC, you need to set some environment variables. Typically,
 * Add `$RF4A_DIR` to the `$PERL5LIB` environment variable:
 
 		export PERL5LIB="$PERL5LIB:$RF4A_DIR"
- 		
+
 * Add `$RF4A_DIR/bin` to your `$PATH` environment variable:
 
 		export PATH="$PATH:$RF4A_DIR/bin"	  
@@ -212,7 +212,7 @@ The following keys are defined:
 
 ### To run the NIST test suite
 
-We use the [NIST FORTRAN78 test suite](ftp://ftp.fortran-2000.com/fcvs21_f95.tar.bz2) for validation. 
+We use the [NIST FORTRAN78 test suite](ftp://ftp.fortran-2000.com/fcvs21_f95.tar.bz2) for validation.
 
       $ cd tests/NIST_F78_test_suite/fcvs21_f95
 
@@ -271,42 +271,41 @@ The final output should look like:
 
 ### To run the Fortran to parallel OpenCL test
 
-This an example of automatic parallelisation and GPU-acceleration of legacy code.  It is the 2-D Shallow Water example from the book "Ocean Modelling for Beginners: Using Open-Source Software" by Jochen Kämpf. The code is effectively Fortran-77 except that it uses Fortran-90 style modules. 
+This an example of automatic parallelisation and GPU-acceleration of legacy code.  It is the 2-D Shallow Water example from the book "Ocean Modelling for Beginners: Using Open-Source Software" by Jochen Kämpf. The code is effectively Fortran-77 except that it uses Fortran-90 style modules.
 
-As explained above, to run this test you need to install the [AutoParallel-Fortran](https://github.com/wimvanderbauwhede/AutoParallel-Fortran) compiler and the [OclWrapper Fortran OpenCL API](https://github.com/wimvanderbauwhede/OpenCLIntegration), as well as [scons](http://scons.org/), a Python-based build system. 
+As explained above, to run this test you need to install the [AutoParallel-Fortran](https://github.com/wimvanderbauwhede/AutoParallel-Fortran) compiler and the [OclWrapper Fortran OpenCL API](https://github.com/wimvanderbauwhede/OpenCLIntegration), as well as [scons](http://scons.org/), a Python-based build system.
 
[…]
