### 2018-08-21 d12bd4e
Many bug fixes as a result of trying to compile the protoms codebase. Currently stuck on include chain for a mixed include.
