+++ README.md
@@ -45,17 +45,17 @@ School of Computing Science, University of Glasgow, UK
 
 * Subroutine extraction  
     * simply add an annotation
-    
+
           !$ACC SUBROUTINE <optional subroutine name>  
           ... <code to be extracted as subroutine>
           !$ACC END SUBROUTINE <optional subroutine name>
-                   
+
 * Call graph generation
 
 ### Automatic parallelisation using OpenCL
 
-Automatic parallelisation and offloading of legacy code to accelerators is the ultimate aim of the project, and already works for many cases. 
-However, the work flow is more complicated and requires an additional compiler, [AutoParallel-Fortran](https://github.com/wimvanderbauwhede/AutoParallel-Fortran). This compiler is written in Haskell, which is not yet a 
+Automatic parallelisation and offloading of legacy code to accelerators is the ultimate aim of the project, and already works for many cases.
+However, the work flow is more complicated and requires an additional compiler, [AutoParallel-Fortran](https://github.com/wimvanderbauwhede/AutoParallel-Fortran). This compiler is written in Haskell, which is not yet a
 mainstream programming language. Furthermore, the generated OpenCL code relies on the [OclWrapper Fortran OpenCL API](https://github.com/wimvanderbauwhede/OpenCLIntegration), written in C++, and uses [scons](http://scons.org/), a Python-based build system. For these reasons, it is harder to install this autoparallelising compiler. However, if you have installed it, a test case for the full flow is provided in the `tests` folder, see below for more details.           
 
 ## How it works
@@ -84,7 +84,7 @@ Furthermore, we tested the compiler on four real-word physics simulation models:
 * The [Linear Baroclinic Model](http://ccsr.aori.u-tokyo.ac.jp/~hiro/sub/lbm.html), an atmospheric climate model (39,336 loc)
 
 
-Each of these models has a different coding style, specifically in terms of the use of common blocks, include files, etc that affect the refactoring process. All of these codes are refactored fully automatically without changes to the original code and build and run correctly. The performance of the original and refactored code is the same in all cases.
+Each of these models has a different coding style, specifically in terms of the use of common blocks, include files, etc. that affect the refactoring process. All of these codes are refactored fully automatically without changes to the original code and build and run correctly. The performance of the original and refactored code is the same in all cases.
 
 ## Installation
 
@@ -101,7 +101,7 @@ To install RefactorF4ACC, you need to set some environment variables. Typically,
 * Add `$RF4A_DIR` to the `$PERL5LIB` environment variable:
 
 		export PERL5LIB="$PERL5LIB:$RF4A_DIR"
- 		
+
 * Add `$RF4A_DIR/bin` to your `$PATH` environment variable:
 
 		export PATH="$PATH:$RF4A_DIR/bin"	  
@@ -139,6 +139,7 @@ The following keys are defined:
 <dt>RENAME_EXT:</dt><dd>Extension for variables that need to be renamed because of conflicts (usually you don't need this; the default is _GLOB)</dd>
 <dt>NO_ONLY:</dt><dd>Do not use the ONLY qualifier on the USE declaration</dd>
 <dt>SPLIT_LONG_LINES:</dt><dd>Split long lines into chunks of no more than 80 characters</dd>
+<dt>MAX_LINE_LENGTH:</dt><dd>Maximum line length for fixed-format F77 code. The default is 132 characters.</dd>
[…]
