### 2018-09-07 wimvanderbauwhede/RefactorF4Acc#2 @rouson
Is there a release version?
Is there a plan to produce numbered releases?  I'm unable to find a version number in Refactor4FAcc.  Release numbers can be especially useful for communicating stability (users might expect releases to be more stable than pre-release commits).  Also, users who want to invoke Refactor4FAcc from a script might like to have the script check the version number using something like
```
RefactorF4Acc --version
```
or
```
RefactorF4Acc -v
```
