### 2018-09-07 wimvanderbauwhede/RefactorF4Acc#3 @rouson
Automated tests
Is there a way to run the test suite automatically?  I tried the following:
```
$ ./NIST_F78_test_suite/fcvs21_f95/Test_rf4a/run_testsuite.pl
cp: ../FM001.f: No such file or directory
TEST CASE FM001.f
refactorF4acc.pl -c rf4a_fm001.cfg  -g fm001
The EQUIVALENCE  statement is not supported, please rewrite your code (or ignore at your peril):
SOURCE: ./NIST_F78_test_suite/fcvs21_f95/FM500.f LINE #34
CODE UNIT: an502
LINE: 'equivalence (ay3s, dz3s)
...
```
and there were hundreds more lines with similar messages.
