### 2018-09-07 #865 @rouson
[issue3 ](https://github.com/wimvanderbauwhede/RefactorF4Acc/issues/3)
