### 2018-09-07 wimvanderbauwhede/RefactorF4Acc#4 @rouson
Contributing guidelines
For third parties wishing to contribute to this project, please document your guidelines, including how to

1) Contribute to the software
2) Report issues or problems with the software
3) Seek support
