### 2018-09-07 #865 @rouson
[issue #4](https://github.com/wimvanderbauwhede/RefactorF4Acc/issues/4)
