### 2018-09-12 wimvanderbauwhede/RefactorF4Acc#2 @wimvanderbauwhede
I will add this. At the moment the version is 1.0.0, I will add a --version flag
