### 2018-09-12 wimvanderbauwhede/RefactorF4Acc#3 @wimvanderbauwhede
The way to run the NIST test suite is explained in the README.md under "To run the NIST test suite", and there is a single script to run it. In short:

$ cd tests/NIST_F78_test_suite/fcvs21_f95/Test_rf4a
$ ./generate_and_run.sh

Please ensure that the environment varianble `$FC` is set to the Fortran compiler you want to use. The above is a summary, the README.md has full details.
