### 2018-09-12 wimvanderbauwhede/RefactorF4Acc#4 @wimvanderbauwhede
OK, will do. I'll make a CONTRIBUTE file. Although this is not really a requirement of Open Source Software, isn't it. And it is very unlikely that anyone else would wish to contribute.
