### 2018-09-12 wimvanderbauwhede/RefactorF4Acc#3 @wimvanderbauwhede
About those "The EQUIVALENCE  statement is not supported" messages, my compiler does not attempt to refactor these statements, but they can break referential transparency, so I felt I had to warn the user. Similar for some other features that I have no intention of supporting. In most cases, the resulting code is just fine.
