### 2018-09-16 wimvanderbauwhede/RefactorF4Acc#4 @labarba
Contributing guidelines are required by JOSS, and listed in the [Review Criteria](https://joss.readthedocs.io/en/latest/review_criteria.html).
