### 2018-09-20 #865 @wimvanderbauwhede
I closed issue #3, I will close the others after I've actually added the CONTRIBUTE file and the version option, in the coming days.
