### 2018-09-21 da9c43b
Added CONTRIBUTE.md for JOSS
