### 2018-09-21 7ab6e49
Added an option to show the version number and a CONTRIBUTE.md file.
