+++ CONTRIBUTE.md
@@ -1,0 +1,28 @@
+# Contributing
+
+You are very welcome to contribute to this project, there are many ways to do so:
+
+## Easy ways to contribute
+
+Here are a few ideas for how you can contribute, even if you are new to compilers, git, or Perl:
+
+- Test the compiler with Fortran code you're using and let me know any issues.
+- Make GitHub issues and contribute to the conversations about how to resolve them.
+- Read issues and pull requests that other people created and contribute to the conversation about how to resolve them.
+- Improve the documentation and the tests.
+- Tell your friends and colleagues about the Glasgow Fortran Source-to-Source Compiler.
+
+## How to contribute to the code
+
+Contributors to the Glasgow Fortran Source-to-Source Compiler use GitHub’s pull requests to add/modify its source code. Here’s an outline of the process:
+
+- Create a GitHub issue and get initial feedback from users and maintainers. If the issue is a bug report, please include the code needed to reproduce the problem.
+- Obtain the latest version of the code: Fork the RefactorF4Acc project to your GitHub account, git clone your fork to your computer.
+- Make some or all of your changes/additions and git commit them to your local repository.
+- Share your changes with us via a pull request: git push your local changes to your GitHub fork, then go to GitHub make a pull request.
+
+Also see GitHub’s Set Up Git and Using Pull Requests.
+
+For developing, you will need Perl v5.10 or later installed. If you want to contribute, please open a pull request, preferably one that addresses a known issue.
+
+You must include documentation and tests for any new or improved code. We can provide help and advice on this after you start the pull request.
