### 2018-09-21 wimvanderbauwhede/RefactorF4Acc#2 @wimvanderbauwhede
I've added a version flag.
