### 2018-09-21 wimvanderbauwhede/RefactorF4Acc#4 @wimvanderbauwhede
I've added a CONTRIBUTE.md file, based on another project from the JOSS site.
