### 2018-09-23 3cd284a
Start of proper docs for the compiler datastructures
