### 2018-10-06 wimvanderbauwhede/RefactorF4Acc#5 @rouson
Release tar ball
Because I can't reopen issue #2, I'm submitting this new issue.  My apologies for the delays.  I'll make every effort to respond faster going forward.   I'll attempt to give more a more complete description of what I think is required here.

On the [review page], the **Review checklist for @rouson** section section contains the following item: "**Version:** Does the release version given match the GitHub release (v1.0.0)?".  I interpret this as implying that

 - [ ] There exists a release version of the software available on this repository's [release page].
 - [ ] There exists an associated release tag (such as `v1.0.0`) as mentioned currently on the [release page].

Please complete these steps and check the above boxes before closing this issue.  I recommend using 3 digits for the version number to facilitate [semantic versioning].

My original suggestion in issue #2 is a nice convenience and not a requirement, but for what it's worth, I still don't see how to find out the release version using the `--version` argument.  My steps are below in case you'd like to support this.  It would be great to be able to execute `refactorF4acc.pl --version` without having to define a configure file.  Also, it would be nice to have `refactorF4acc.pl --help` return one-line descriptions of each possible argument.
```
$ git clone https://github.com/wimvanderbauwhede/RefactorF4Acc
Cloning into 'RefactorF4Acc'...
remote: Enumerating objects: 3582, done.
remote: Total 3582 (delta 0), reused 0 (delta 0), pack-reused 3582
Resolving deltas: 100% (2434/2434), done.
$ cd RefactorF4Acc/
$ export PERL5LIB="$PERL5LIB:${PWD}"
$ export PATH="$PATH:${PWD}/bin"
$ refactorF4acc.pl --version
/home/rouson/RefactorF4Acc/bin/refactorF4acc.pl version 1.0.0 calling Getopt::Std::getopts (version 1.12 [paranoid]),
running under Perl version 5.26.1.
  [Now continuing due to backward compatibility and excessive paranoia.
   See 'perldoc Getopt::Std' about $Getopt::Std::STANDARD_HELP_VERSION.]
No default for toplevel subroutine (TOP) in rf4a.cfg, please specify the toplevel subroutine on command line
```
[review page]: https://github.com/openjournals/joss-reviews/issues/865#issuecomment-421817616
[release page]: https://github.com/wimvanderbauwhede/RefactorF4Acc/releases
[semantic versioning]: https://semver.org
