### 2018-10-06 #865 @rouson
I've added [issue 5] to better explain what I think was required for issue 2, which I don't think I have the ability to reopen.

[issue 5]: https://github.com/wimvanderbauwhede/RefactorF4Acc/issues/5
