### 2018-10-07 wimvanderbauwhede/RefactorF4Acc#6 @rouson
Short example
It would be instructive to provide an example that succinctly demonstrates as many of the capabilities in the [Features] list as is feasible in 50 to 100 lines of code.

[Features]: https://github.com/wimvanderbauwhede/RefactorF4Acc/blob/master/README.md#features
