### 2018-10-07 #865 @rouson
I have now run the NIST test suite using the [directions] in the README.md.  I would feel more confident my verification of this software's functionality if the repository contained a much shorter example (say 50-100 lines) that demonstrates the results of the transformations this software provides.  I have created [issue 6] to request this.

[directions]: https://github.com/wimvanderbauwhede/RefactorF4Acc#to-run-the-nist-test-suite
[issue 6]: https://github.com/wimvanderbauwhede/RefactorF4Acc/issues/6
