### 2018-10-07 #865 @rouson
I have submitted [issue 6] got request that a DOI be added to one article reference.  Once issues 5, 6, and 7 have been addressed, I expect to be able to recommend acceptance.
