### 2018-10-12 wimvanderbauwhede/RefactorF4Acc#5 @wimvanderbauwhede
I'm sorry, I used -V for the version flag. If you insist I will change the code to support long flags.
