+++ paper.bib
@@ -3,6 +3,7 @@
   author={Vanderbauwhede, Wim and Davidson, Gavin},
   journal={Computers \& Fluids},
   year={2018},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={https://doi.org/10.1016/j.compfluid.2018.06.005}
 }
 
