### 2018-10-12 a1b0dc7
Added a README to the 2-D shallow water example, changed the path in the Auto-acceleration-README.md, changed SConstruct.rf4a to use $FC
