+++ README.md
@@ -41,7 +41,7 @@ School of Computing Science, University of Glasgow, UK
   * Preserves comments
 
 * OpenCL/C translation
-  * Once refactored, modules can be translated to C or OpenCL kernel code in a separate pass
+  * Once refactored, modules can be translated to C or OpenCL kernel code in a separate pass (see [Automatic parallelisation using OpenCL](https://github.com/wimvanderbauwhede/RefactorF4Acc/blob/master/README.md#automatic-parallelisation-using-opencl)) 
 
 * Subroutine extraction  
     * simply add an annotation
