### 2018-10-12 wimvanderbauwhede/RefactorF4Acc#6 @wimvanderbauwhede
There is an example of the 2-D Shallow Water equation [in the tests folder](https://github.com/wimvanderbauwhede/RefactorF4Acc/tree/master/tests/ShallowWater2D), does that not meet your requirements? It is 180 lines of code in total. Although this code has .f95 extensions, it is actually F77 code with F95 type declarations. It is taken from the book "Ocean Modelling for Beginners" by Kaempf.

The only feature this example does not demonstrate is GOTO conversion, but there are plenty of these in the NIST testbench. I have added a README.md.
