### 2018-10-12 39c0f1e
Now with --version and --help
