### 2018-10-12 wimvanderbauwhede/RefactorF4Acc#5 @wimvanderbauwhede
OK, I've added --version and --help.
