### 2018-10-21 #865 @labarba
I just pinged @rasmussn by email to ask if he can move forward with this review soon.

@wimvanderbauwhede : Let us know here when you address the issues raised by @rouson.
