### 2018-10-22 #865 @wimvanderbauwhede
I have addressed all issues but  I am waiting for a reply from @rouson on
[issue 6](https://github.com/wimvanderbauwhede/RefactorF4Acc/issues/6).
