### 2018-10-22 #865 @rouson
I just added a comment on [issue 6](
https://github.com/wimvanderbauwhede/RefactorF4Acc/issues/6).
