### 2018-10-23 wimvanderbauwhede/RefactorF4Acc#8 @rouson
Results of a "Hello,  world" attempt
I am very interested in using RefactorF4Acc on the [zgoubi] project, a ~300,000-line beam physics code developed at Brookhaven National Laboratory in the U.S.  I haven't had success in running `refactorF4Acc.pl` so I wrote a 26-line "Hello, world!" program ([hello.tar.gz]) that exhibits a few of the important features of [zgoubi], including implicit typing, a `common` block, fixed-form source, and subroutine dummy arguments with unspecified `intent`.  I'm attaching the "Hello, world!" program and providing the transcript below that shows how to build and run the code as well as a depiction of the directory structure and the `refactorF4Acc.pl` error message.

It would be great to include a working version of this "Hello, world!" or something similar in the RefactorF4Acc tests subdirectory as a way for users to get started.

```bash
$ tar xf hello.tar.gz
$ cd hello
$ make
gfortran -o hello src/main.f src/greet_world.f -I src/include
$ ./hello
 Hello, world!
 All is right with the world.
$ tree
.
├── Makefile
├── rf4a.cfg
└── src
    ├── greet_world.f
    ├── include
    │   └── error_codes.h
    └── main.f

2 directories, 5 files
$ export PERL5LIB="$PERL5LIB:${HOME}/Builds/RefactorF4Acc"
$ export PATH="$PATH:${HOME}/Builds/RefactorF4Acc/bin"
$ cat rf4a.cfg
TOP = hello
PREFIX = .
SRCDIRS = src
NEWSRCPATH = refactored-source
$ refactorF4acc.pl -c rf4a.cfg
Expected ClosingParen, got EOF at /home/rouson/Builds/RefactorF4Acc/Fortran/Expression/Evaluator/Parser.pm line 201.
	Fortran::Expression::Evaluator::Parser::_match(Fortran::Expression::Evaluator::Parser=HASH(0x558dede18408), "ClosingParen") called at /home/rouson/Builds/RefactorF4Acc/Fortran/Expression/Evaluator/Parser.pm line 244
	Fortran::Expression::Evaluator::Parser::_function_call(Fortran::Expression::Evaluator::Parser=HASH(0x558dede18408)) called at /home/rouson/Builds/RefactorF4Acc/Fortran/Expression/Evaluator/Parser.pm line 217
	Fortran::Expression::Evaluator::Parser::_value(Fortran::Expression::Evaluator::Parser=HASH(0x558dede18408)) called at /home/rouson/Builds/RefactorF4Acc/Fortran/Expression/Evaluator/Parser.pm line 360
	Fortran::Expression::Evaluator::Parser::_factor(Fortran::Expression::Evaluator::Parser=HASH(0x558dede18408)) called at /home/rouson/Builds/RefactorF4Acc/Fortran/Expression/Evaluator/Parser.pm line 369
	Fortran::Expression::Evaluator::Parser::_exponential(Fortran::Expression::Evaluator::Parser=HASH(0x558dede18408)) called at /home/rouson/Builds/RefactorF4Acc/Fortran/Expression/Evaluator/Parser.pm line 302
	Fortran::Expression::Evaluator::Parser::_term(Fortran::Expression::Evaluator::Parser=HASH(0x558dede18408)) called at /home/rouson/Builds/RefactorF4Acc/Fortran/Expression/Evaluator/Parser.pm line 336
	Fortran::Expression::Evaluator::Parser::_expression(Fortran::Expression::Evaluator::Parser=HASH(0x558dede18408)) called at /home/rouson/Builds/RefactorF4Acc/Fortran/Expression/Evaluator/Parser.pm line 260
	Fortran::Expression::Evaluator::Parser::_statement(Fortran::Expression::Evaluator::Parser=HASH(0x558dede18408)) called at /home/rouson/Builds/RefactorF4Acc/Fortran/Expression/Evaluator/Parser.pm line 175
	Fortran::Expression::Evaluator::Parser::_program(Fortran::Expression::Evaluator::Parser=HASH(0x558dede18408)) called at /home/rouson/Builds/RefactorF4Acc/Fortran/Expression/Evaluator/Parser.pm line 128
	Fortran::Expression::Evaluator::Parser::parse("character(len", HASH(0x558dede18750)) called at /home/rouson/Builds/RefactorF4Acc/RefactorF4Acc/Parser/Expressions.pm line 171
	RefactorF4Acc::Parser::Expressions::parse_expression("character(len", HASH(0x558decbb0690), HASH(0x558decbade30), "hello") called at /home/rouson/Builds/RefactorF4Acc/RefactorF4Acc/Parser.pm line 4245
	RefactorF4Acc::Parser::_parse_assignment("character(len=5)greeting", HASH(0x558decbb0690), HASH(0x558decbade30), "hello") called at /home/rouson/Builds/RefactorF4Acc/RefactorF4Acc/Parser.pm line 1257
	RefactorF4Acc::Parser::_analyse_lines("hello", HASH(0x558decbade30)) called at /home/rouson/Builds/RefactorF4Acc/RefactorF4Acc/Parser.pm line 99
	RefactorF4Acc::Parser::parse_fortran_src("hello", HASH(0x558decbade30)) called at /home/rouson/Builds/RefactorF4Acc/bin/refactorF4acc.pl line 168
	main::main() called at /home/rouson/Builds/RefactorF4Acc/bin/refactorF4acc.pl line 67

```

[hello.tar.gz]: https://github.com/wimvanderbauwhede/RefactorF4Acc/files/2508399/hello.tar.gz

[zgoubi]: https://github.com/radiasoft/zgoubi
