### 2018-10-24 wimvanderbauwhede/RefactorF4Acc#8 @rouson
Also, I just tried simplifying this example down to two lines of Fortran source code and found that `refactorF4acc.pl` fails without the `-g` flag.  Is this flag required for all codes?
```bash
$ ls
main.f  rf4a.cfg
$ cat rf4a.cfg
TOP = hello
PREFIX = .
SRCDIRS = .
NEWSRCPATH = refactored-source
$ cat main.f
      PROGRAM hello
      END PROGRAM
$ refactorF4acc.pl -c rf4a.cfg
SUB hello DOES NOT HAVE RefactoredArgs
BOOM! 271 hello : $VAR1 = {
  'FreeForm' => 0,
  'Entries' => {
    'Set' => {},
    'List' => []
  },
  'UndeclaredCommonVars' => {
    'Set' => {},
    'List' => []
  },
  'DoneInitTables' => 1,
  'UsedGlobalVars' => {
    'List' => [],
    'Set' => {}
  },
  'ExGlobArgs' => {
    'Set' => {},
    'List' => []
  },
  'DeclaredOrigArgs' => {
    'Set' => {},
    'List' => []
  },
  'Entry' => 0,
  'HasRefactoredArgs' => 0,
  'RefactorGlobals' => 0,
  'DeclaredOrigLocalVars' => {
    'Set' => {},
    'List' => []
  },
  'LocalParameters' => {
    'List' => [],
    'Set' => {}
  },
  'Callers' => {},
  'UndeclaredOrigArgs' => {
    'Set' => {},
    'List' => []
  },
  'UndeclaredOrigLocalVars' => {
    'Set' => {},
    'List' => []
  },
  'Recursive' => 0,
  'RefactoredCode' => [
    [
      '      program hello',
      {
        'Indent' => '      ',
        'Ref' => 0,
        'Signature' => {
          'Name' => 'hello',
          'Program' => 1,
          'Args' => {
            'List' => []
          }
        },
        'Block' => {
          'Nest' => 1,
          'InBlock' => {},
          'Type' => 'program',
          'LineID' => 0,
          'Name' => 'hello'
        },
        'LineID' => 0,
        'AccPragma' => {},
        'SubroutineSig' => [
          '      ',
          'hello',
          []
        ],
        'ExGlobVarDeclHook' => 'Signature'
      }
    ],
    [
      '      end program',
      {
        'ExtraIncludesHook' => 1,
        'ExtraModulesHook' => 1,
        'LineID' => 1,
        'Block' => $VAR1->{'RefactoredCode'}[0][1]{'Block'},
        'Indent' => '      ',
        'Ref' => 0,
        'AccPragma' => {}
      }
    ]
  ],
  'UsedParameters' => {
    'Set' => {},
    'List' => []
  },
  'TabFormat' => 0,
  'Vars' => {
    'Subsets' => {
      'CommonVars' => {
        'Subsets' => {
          'UndeclaredCommonVars' => $VAR1->{'UndeclaredCommonVars'},
          'DeclaredCommonVars' => {
            'List' => [],
            'Set' => {}
          }
        }
      },
      'Args' => {
        'Subsets' => {
          'ExGlobArgs' => $VAR1->{'ExGlobArgs'},
          'ExInclArgs' => {
            'Set' => {},
            'List' => []
          },
          'OrigArgs' => {
            'List' => [],
            'Subsets' => {
              'UndeclaredOrigArgs' => $VAR1->{'UndeclaredOrigArgs'},
              'DeclaredOrigArgs' => $VAR1->{'DeclaredOrigArgs'}
            }
          }
        }
      },
      'Parameters' => {
        'Subsets' => {
          'UsedParameters' => $VAR1->{'UsedParameters'},
          'LocalParameters' => $VAR1->{'LocalParameters'},
          'IncludedParameters' => {
            'Set' => {},
            'List' => []
          },
          'ParametersFromContainer' => {
            'List' => [],
            'Set' => {}
          }
        }
      },
      'LocalVars' => {
        'Subsets' => {
          'OrigLocalVars' => {
            'Subsets' => {
              'DeclaredOrigLocalVars' => $VAR1->{'DeclaredOrigLocalVars'},
              'UndeclaredOrigLocalVars' => $VAR1->{'UndeclaredOrigLocalVars'}
            }
          },
          'UsedLocalVars' => {
            'Set' => {},
            'List' => []
          },
          'ExInclLocalVars' => {
            'List' => [],
            'Set' => {}
          }
        }
      }
    }
  },
  'OrigLocalVars' => $VAR1->{'Vars'}{'Subsets'}{'LocalVars'}{'Subsets'}{'OrigLocalVars'},
  'RefactoredArgs' => {
    'List' => [],
    'Set' => {}
  },
  'ExGlobVarDeclHook' => 1,
  'MaskedIntrinsics' => {},
  'ConflictingParams' => {},
  'Source' => './main.f',
  'FStyle' => 'F77',
  'Program' => 1,
  'Name' => 'hello',
  'HasType' => '',
  'UsedLocalVars' => $VAR1->{'Vars'}{'Subsets'}{'LocalVars'}{'Subsets'}{'UsedLocalVars'},
  'AnnLines' => $VAR1->{'RefactoredCode'},
  'LocalVars' => $VAR1->{'Vars'}{'Subsets'}{'LocalVars'},
  'ExInclLocalVars' => $VAR1->{'Vars'}{'Subsets'}{'LocalVars'}{'Subsets'}{'ExInclLocalVars'},
  'ParametersFromContainer' => $VAR1->{'Vars'}{'Subsets'}{'Parameters'}{'Subsets'}{'ParametersFromContainer'},
  'IncludedParameters' => $VAR1->{'Vars'}{'Subsets'}{'Parameters'}{'Subsets'}{'IncludedParameters'},
  'OrigArgs' => $VAR1->{'Vars'}{'Subsets'}{'Args'}{'Subsets'}{'OrigArgs'},
  'ExInclArgs' => $VAR1->{'Vars'}{'Subsets'}{'Args'}{'Subsets'}{'ExInclArgs'},
  'ReferencedLabels' => {},
  'Parameters' => $VAR1->{'Vars'}{'Subsets'}{'Parameters'},
  'Elemental' => 0,
  'Args' => $VAR1->{'Vars'}{'Subsets'}{'Args'},
  'RenamedInheritedExGLobs' => {
    'Set' => {}
  },
  'HasBlocks' => 0,
  'Status' => 3,
  'InheritedParameters' => {
    'List' => [],
    'Set' => {}
  },
  'Commons' => {},
  'Includes' => {},
  'Pure' => 0,
  'CalledEntries' => {
    'Set' => {},
    'List' => []
  },
  'CalledSubs' => {
    'Set' => {},
    'List' => []
  },
  'HasCommons' => 0,
  'CommonVars' => $VAR1->{'Vars'}{'Subsets'}{'CommonVars'},
  'RenamedInheritedExGlobs' => {
    'List' => [],
    'Set' => {}
  },
  'DeclaredCommonVars' => $VAR1->{'Vars'}{'Subsets'}{'CommonVars'}{'Subsets'}{'DeclaredCommonVars'}
};
 at /home/rouson/Desktop/Builds/RefactorF4Acc/RefactorF4Acc/Analysis/ArgumentIODirs.pm line 271.
	RefactorF4Acc::Analysis::ArgumentIODirs::_analyse_src_for_iodirs(HASH(0x555b98b3eef0), "hello") called at /home/rouson/Desktop/Builds/RefactorF4Acc/RefactorF4Acc/Analysis/ArgumentIODirs.pm line 77
	RefactorF4Acc::Analysis::ArgumentIODirs::_determine_argument_io_direction_core(HASH(0x555b98b3eef0), "hello") called at /home/rouson/Desktop/Builds/RefactorF4Acc/RefactorF4Acc/Analysis/ArgumentIODirs.pm line 60
	RefactorF4Acc::Analysis::ArgumentIODirs::determine_argument_io_direction_rec(HASH(0x555b98b3eef0), "hello") called at /home/rouson/Desktop/Builds/RefactorF4Acc/RefactorF4Acc/Refactoring.pm line 83
	RefactorF4Acc::Refactoring::refactor_all(HASH(0x555b98b3eef0), "hello", "") called at /home/rouson/Desktop/Builds/RefactorF4Acc/bin/refactorF4acc.pl line 192
	main::main() called at /home/rouson/Desktop/Builds/RefactorF4Acc/bin/refactorF4acc.pl line 67
```
