### 2018-10-24 wimvanderbauwhede/RefactorF4Acc#8 @rouson
In case it's helpful, I also reduced the first example above to one that exhibits what appears to be the same failure even with the `-g` flag and after cutting out ~20 lines of code:
```bash
$ cat rf4a.cfg
TOP = hello
PREFIX = .
SRCDIRS = .
NEWSRCPATH = refactored-source
$ cat main.f
```
```fortran
      subroutine greet_world(string)
        character(len=*) string
        write(*,*) string
      end subroutine

      program hello
        implicit none
        call greet_world("Hello, world!")
      end program
```
```bash
$ refactorF4acc.pl -c rf4a.cfg -g
Expected ClosingParen, got EOF at /home/rouson/Desktop/Builds/RefactorF4Acc/Fortran/Expression/Evaluator/Parser.pm line 201.
	Fortran::Expression::Evaluator::Parser::_match(Fortran::Expression::Evaluator::Parser=HASH(0x556d8fb4e6f0), "ClosingParen") called at /home/rouson/Desktop/Builds/RefactorF4Acc/Fortran/Expression/Evaluator/Parser.pm line 244
	Fortran::Expression::Evaluator::Parser::_function_call(Fortran::Expression::Evaluator::Parser=HASH(0x556d8fb4e6f0)) called at /home/rouson/Desktop/Builds/RefactorF4Acc/Fortran/Expression/Evaluator/Parser.pm line 217
	Fortran::Expression::Evaluator::Parser::_value(Fortran::Expression::Evaluator::Parser=HASH(0x556d8fb4e6f0)) called at /home/rouson/Desktop/Builds/RefactorF4Acc/Fortran/Expression/Evaluator/Parser.pm line 360
	Fortran::Expression::Evaluator::Parser::_factor(Fortran::Expression::Evaluator::Parser=HASH(0x556d8fb4e6f0)) called at /home/rouson/Desktop/Builds/RefactorF4Acc/Fortran/Expression/Evaluator/Parser.pm line 369
	Fortran::Expression::Evaluator::Parser::_exponential(Fortran::Expression::Evaluator::Parser=HASH(0x556d8fb4e6f0)) called at /home/rouson/Desktop/Builds/RefactorF4Acc/Fortran/Expression/Evaluator/Parser.pm line 302
	Fortran::Expression::Evaluator::Parser::_term(Fortran::Expression::Evaluator::Parser=HASH(0x556d8fb4e6f0)) called at /home/rouson/Desktop/Builds/RefactorF4Acc/Fortran/Expression/Evaluator/Parser.pm line 336
	Fortran::Expression::Evaluator::Parser::_expression(Fortran::Expression::Evaluator::Parser=HASH(0x556d8fb4e6f0)) called at /home/rouson/Desktop/Builds/RefactorF4Acc/Fortran/Expression/Evaluator/Parser.pm line 260
	Fortran::Expression::Evaluator::Parser::_statement(Fortran::Expression::Evaluator::Parser=HASH(0x556d8fb4e6f0)) called at /home/rouson/Desktop/Builds/RefactorF4Acc/Fortran/Expression/Evaluator/Parser.pm line 175
	Fortran::Expression::Evaluator::Parser::_program(Fortran::Expression::Evaluator::Parser=HASH(0x556d8fb4e6f0)) called at /home/rouson/Desktop/Builds/RefactorF4Acc/Fortran/Expression/Evaluator/Parser.pm line 128
	Fortran::Expression::Evaluator::Parser::parse("character(len", HASH(0x556d8fb4e6c0)) called at /home/rouson/Desktop/Builds/RefactorF4Acc/RefactorF4Acc/Parser/Expressions.pm line 171
	RefactorF4Acc::Parser::Expressions::parse_expression("character(len", HASH(0x556d8fb3c920), HASH(0x556d8e8cbe30), "greet_world") called at /home/rouson/Desktop/Builds/RefactorF4Acc/RefactorF4Acc/Parser.pm line 4245
	RefactorF4Acc::Parser::_parse_assignment("character(len=*)string", HASH(0x556d8fb3c920), HASH(0x556d8e8cbe30), "greet_world") called at /home/rouson/Desktop/Builds/RefactorF4Acc/RefactorF4Acc/Parser.pm line 1257
	RefactorF4Acc::Parser::_analyse_lines("greet_world", HASH(0x556d8e8cbe30)) called at /home/rouson/Desktop/Builds/RefactorF4Acc/RefactorF4Acc/Parser.pm line 99
	RefactorF4Acc::Parser::parse_fortran_src("greet_world", HASH(0x556d8e8cbe30)) called at /home/rouson/Desktop/Builds/RefactorF4Acc/RefactorF4Acc/Parser.pm line 1902
	RefactorF4Acc::Parser::_parse_subroutine_and_function_calls("hello", HASH(0x556d8e8cbe30)) called at /home/rouson/Desktop/Builds/RefactorF4Acc/RefactorF4Acc/Parser.pm line 118
	RefactorF4Acc::Parser::parse_fortran_src("hello", HASH(0x556d8e8cbe30)) called at /home/rouson/Desktop/Builds/RefactorF4Acc/bin/refactorF4acc.pl line 168
	main::main() called at /home/rouson/Desktop/Builds/RefactorF4Acc/bin/refactorF4acc.pl line 67
```
