### 2018-10-24 #865 @rouson
I've also added [issue 8] that probably needs to be addressed before acceptance.  For new users, I think the documentation should contain a working version of an extremely simple example similar to one of the "Hello, world!" examples in [issue 8].

 [issue 8]: https://github.com/wimvanderbauwhede/RefactorF4Acc/issues/8
