### 2018-10-24 wimvanderbauwhede/RefactorF4Acc#6 @wimvanderbauwhede
Thanks for the comments.
(1) You are right of course, I was simplifying a bit when I said the code only used F95 type declarations. I am familiar with the F77 standard :-) I have included the differences you list in the README.md

(2) The SCons error is because you didn't have $CC defined. You don't need it so I removed it. I also made a script "generate.sh" which does not build the code, so it does not need SCons. And I added the manual build line to the README.md

(3) I use cloc to count lines of code, and it gives

    [wim@HackBookPro fortran]$ cloc --include-lang='Fortran 95' .
           7 text files.
           7 unique files.
           4 files ignored.

    Language                     files          blank        comment           code
    Fortran 95                       3             74             81            180
    SUM:                             3             74             81            180

I would appreciate if you would submit a pull request for a shorter version of the code. I did the obvious things and it is now

    [wim@HackBookPro fortran]$ wc -l *.f95
         119 main.f95
          19 param.f95
         149 sub.f95
         287 total

I'm not sure what the difference would be between GETTING_STARTED.md and README.md, and also not really about what changes you'd like to be described in the README.md, could you explain?

Thanks!
