### 2018-10-24 ccc9ddf
Updated README.md with info on how to build without SCons, added generate.sh which and removed redundant lines from the source.
