### 2018-10-24 22d2774
Add Makefile for refactored shallow-water source
This won't be used in the generate*.sh scripts so that those
scripts depend only on the presence of Refactor4FAcc and gfortran.
But having the Makefile is useful for subsequent rebuilds and
especially for "make clean".
