### 2018-10-24 1316a3e
Eliminate scons requirement in shallow-water build
