### 2018-10-24 wimvanderbauwhede/RefactorF4Acc#6 @rouson
The changes look great!   I believe the code changes suffice so there's no need for me to submit a pull request with simplified code.  I'll close this issue now and will submit the other related pull requests with a GETTING_STARTED.md and a few other suggested documentation changes.
