### 2018-10-24 #865 @rouson
I have closed [issue 6].  I expect to recommend acceptance after [issue 8] has been addressed.

[issue 6]:
https://github.com/wimvanderbauwhede/RefactorF4Acc/issues/6

[issue 8]:
https://github.com/wimvanderbauwhede/RefactorF4Acc/issues/8
