### 2018-10-24 wimvanderbauwhede/RefactorF4Acc#8 @wimvanderbauwhede
Hi Damian,

Thank you for your efforts! I had fixed the parse error you got in my devel branch but I have not yet merged that one into master.  As soon as I run the regression tests and merged, I will add the Hello, World examples you've provided.

The -g flag is currently necessary so I changed it into a no-op.

It would be nice if my compiler could handle your 300K codebase. It depends more on the code features than on the size of the code. At the moment I don't really have time to test it, and that's a pity as I am quite keen to do this.
