### 2018-10-24 #865 @rouson
[Pull request 10] has been submitted with severals suggested documentation improvements.   As soon as the parser fix has been pushed, I'll work on a second pull request to resolve [issue 8], which is the last issue related to this review.  The second pull request will include the GETTING_STARTED.md.

[Pull request 10]: https://github.com/wimvanderbauwhede/RefactorF4Acc/pull/10
[issue 8]:
https://github.com/wimvanderbauwhede/RefactorF4Acc/issues/8
