### 2018-10-25 c2cf1cb
Added README.md for NIST tests, ran regression tests and changed version to 1.1.0
