+++ tests/NIST_F78_test_suite/fcvs21_f95/RefactoredSources/extract_overall_performance.sh
@@ -1,5 +1,5 @@
 grep 'REQUIRE INSPEC' *.res | grep -v '0 TESTS' > REQUIRE_INSPECTION.txt 
 grep EXECUTED *.res > EXECUTED.txt 
-grep FAILED *.res | grep -v '0 TESTS' > FAIL.txt 
+grep 'TESTS FAILED' *.res | grep -v '0 TESTS' > FAIL.txt 
 grep PASSED *.res | grep -v '0 TESTS' > PASS.txt
 ./get_totals.pl
