### 2018-10-31 wimvanderbauwhede/RefactorF4Acc#8 @wimvanderbauwhede
I have just merged the devel branch so the "Hello, World!" example is now working. I've also added a generate shell script and a Makefile for the generated code.
