### 2018-10-31 wimvanderbauwhede/RefactorF4Acc#8 @wimvanderbauwhede
I also verified the NIST tests and the Shallow Water, they all work so I'm closing this issue.
