### 2018-10-31 #865 @wimvanderbauwhede
I think I have now closed all remaining issues. Thank you very much for all your comments and inputs.
