### 2018-10-31 #865 @labarba
@rouson -- Do recommend acceptance, at this point?

I have contacted @rasmussn by email to ask about his review, as he's been rather quiet here. Stay tuned.
