### 2018-10-31 #865 @wimvanderbauwhede
I'm sorry, I don't understand what you mean by "not resolving properly". I assume this is about the DOI in [paper.bib](https://github.com/wimvanderbauwhede/RefactorF4Acc/blob/master/paper.bib)?

The DOI is [https://doi.org/10.1016/j.compfluid.2018.06.005](https://doi.org/10.1016/j.compfluid.2018.06.005) and it resolves to [https://www.sciencedirect.com/science/article/pii/S0045793018302950?via%3Dihub](https://www.sciencedirect.com/science/article/pii/S0045793018302950?via%3Dihub) and I get the full paper.  I've tried this from home as well as from the University network. What happens when you try this?
