### 2018-10-31 wimvanderbauwhede/RefactorF4Acc#8 @rouson
Awesome!  RefactorF4Acc rocks!  I love what it does to the "Hello, world!" example.  I'll begin investigating whether I can use RefactorF4Acc on [zgoubi](https://github.com/radiasoft/zgoubi).  I will submit a pull request with a `GETTING_STARTED.md` that highlights describes the "Hello, world!" example.
