### 2018-10-31 #865 @rouson
@labarba I just finished checking all boxes on my reviewer checklist.  All issues that I submitted as part of this review have been resolved and closed.  I enthusiastically recommend acceptance.  Thanks for the invitation to review.

@wimvanderbauwhede Thanks for the great contribution.  There's a reasonable chance this software will positively impact one of my current projects and I'll definitely use it in teaching, including in a [course] I'm teaching next week.

[course]: http://writing-fortran-2018-today.eventbrite.com
