+++ paper.bib
@@ -4,6 +4,7 @@
   journal={Computers \& Fluids},
   year={2018},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/j.compfluid.2018.06.005}
+  url={https://doi.org/10.1016/j.compfluid.2018.06.005},
+  doi={10.1016/j.compfluid.2018.06.005}
 }
 
