### 2018-10-31 #865 @wimvanderbauwhede
OK, I changed the DOI in paper.bib, I hope it works now.
