### 2018-10-31 #865 @labarba
The DOI link works now!

I've pinged @rasmussn by email again. I hope we can get this finished in the next few days.
