### 2018-10-31 69d4870
Add GETTING_STARTED.md
