### 2018-10-31 f36c431
Update GETTING_STARTED.md
