+++ GETTING_STARTED.md
@@ -1,13 +1,13 @@
 #  Refactoring "Hello, world!"
-The following command-line sessions transcript demonstrates the use of RefactorF4Acc to 
+The following command-line session transcript demonstrates the use of RefactorF4Acc to 
 refactor a "Hello, world!" program that starts with the following legacy Fortran features:
-* Fixed-form source
-* A Fortran 77 parameter statement
-* An implicitly typed variable
-* A `common` block
-* An `include` statement
-* Dummy arguments with no specified `intent`
-* An external subroutine with an implicit interface
+* Fixed-form source as designated by the `.f` file extensions
+* A Fortran 77 parameter statement.
+* An implicitly typed variable (`isuccess`).
+* A `common` block in `error_codes.h`.
+* An `include` statement in `main.f`.
+* Dummy arguments with no specified `intent`.
+* An external subroutine with an implicit interface.
 ```
 # Display source directory tree structure
 $ tree .
@@ -55,7 +55,20 @@ $ cat src/include/error_codes.h
 C implicitly typed global data
       common/error_codes/ isuccess
 ```
-RefactorF4Acc can refactor the above code using using the following configure file:
+With the above directory tree structure, the following Makefile builds the code:
+```
+$ cat Makefile
+hello: src/main.f
+	gfortran -o hello src/main.f src/greet_world.f -I src/include
+
+.PHONY : clean
+
+clean : 
+	rm hello
+```
+Compiling and executing the original source can be accomplished by executing `make && ./hello`.
+
+RefactorF4Acc can refactor the above code using using the following configure file and command:
 ```
 $ cat rf4a.cfg 
 TOP = hello
@@ -63,6 +76,15 @@ PREFIX = .
 SRCDIRS = src
 NEWSRCPATH = refactored-source
 $ refactorF4acc.pl -c rf4a.cfg 
+```
+The resulting code has the following properties:
+* Free form source as indicated by the new `.f95` file extensions.
+* Use association of a module procedure, giving `greet_world` an explicit interface to improve type safety.
+* Specified dummy argument `intent` inferred by source analysis.
+* A module variable replacing the `common` block.
+* `!` comments
+The resulting new directory tree structure is as follows:
+```
 $ tree refactored-source/
 refactored-source/
 └── src
@@ -70,17 +92,9 @@ refactored-source/
     ├── include
     │   └── error_codes.f95
     └── main.f95
-
 ```
-The resulting code has the following properties:
-* Free form source as indicated by the new `.f95` file extensions.
-* Use association of a module procedure (which gives the procedure an explicit interface, thereby improving type safety).
-* Specified dummy argument `intent` inferred by source analysis.
-* A module variable replacing the `common` block.
-* `!` comments
 Because the comments all related to legacy features that RefactorF4Acc removes or transforms,  the comments 
 were removed manually to avoid confusion in the following listings:
-related to 
 ```fortran
 $ cat refactored-source/src/main.f95 
       program hello
