### 2018-10-31 7f3804f
Add link to INSTALL.md in GETTING_STARTED.md
