### 2018-10-31 8e168c5
Eliminate the Makefile from GETTING_STARTED.md
