### 2018-10-31 6242973
Documentation updates
README.md updates:
* Move installation information higher.
* Replace "Usage" section with new "Getting Started" section,
  which includes links to GETTING_STARTED.md

INSTALL updates
* Renamed to INSTALL.md
* Markdown-formatted header
