### 2018-11-04 da83b84
Renamed IdentifyStencils.pm to ArrayAccessPatterns.pm; made .f90 the default extension; added a VERSIONING file; fixed the rename_array_accesses_to_scalars pass. There is still the problem that _analyse_src_for_iodirs only works on RefactoredArgs, I should make sure all args are grouped in RefactoredArgs.
