+++ tests/ShallowWater2D/fortran/generate_and_build.sh
@@ -1,3 +1,3 @@
-refactorF4acc.pl $* -c ./rf4a.cfg -g
+refactorF4acc.pl $* -c ./rf4a.cfg -b
 cd ../RefactoredSources
 scons -f SConstruct.rf4a
