+++ tests/hello/rf4a.cfg
@@ -1,0 +1,4 @@
+TOP = hello
+PREFIX = .
+SRCDIRS = src
+NEWSRCPATH = refactored-source
