### 2018-11-07 b4d1831
Trying to document the compiler datastructures
