### 2018-11-09 d519715
TyTraCL rewrite system to remove redundant memory allocation.
