+++ TyTraCLTransformations/memory-reduction-TyTraCL.hs
@@ -1,0 +1,208 @@
+{-# LANGUAGE DeriveDataTypeable #-}
+module Main where
+
+--import Data.Generics (mkQ, mkT, gmapQ, gmapT, everything, everywhere)
+import Data.Generics (Data, Typeable, mkQ, mkT, gmapQ, gmapT, everything, everywhere, everywhere')
+
+main = do
+    putStrLn "Original AST"
+    mapM print ast
+    putStrLn "\nSplit LHS tuples"
+    mapM print ast'
+    putStrLn "\nSubstitute vectors (recursive)"
+    mapM print ast''
+    putStrLn "\nTest for Vec in RHS Expr"
+    mapM (print . get_vec_subexprs . snd) ast''
+    putStrLn "\nTest for non-input Vec in RHS Expr"
+    mapM (print . expr_has_non_input_vecs . snd) ast'
+    putStrLn "\nTest for non-input Vec in RHS Expr"
+    mapM (print . expr_has_non_input_vecs . snd) ast''
+--    print $ find_in_ast ast' (Vec VT "duu_1")
+
+{-
+    main (etan_0,wet_0,eps_0,dt_0,g_0,dx_0,dy_0,wet_0,u_0,v_0,duu_0,dvv_0,h_0,dt_0,dx_0,dy_0,hzero_0,hmin_0) =
+      let
+         -- shapiro_map_15
+        s1 = [-1,-502,0,502,1]
+        etan_s_0 = stencil s1 etan_0
+        s2 = [-1,-502,0,502,1]
+        wet_s_0 = stencil s2 wet_0
+        eta_1 =  map (shapiro_map_15 eps_0) (zipt (wet_s_0,etan_s_0))
+         -- dyn_map_38
+        s3 = [0,502,1]
+        eta_s_0 = stencil s3 eta_1
+        (du_1,dv_1) = unzipt $ map (dyn_map_38 (dt_0,g_0,dx_0,dy_0)) eta_s_0
+         -- dyn_map_44
+        s4 = [0,502,1]
+        wet_s_1 = stencil s4 wet_0
+        (duu_1,dvv_1,un_1,vn_1) = unzipt $ map (dyn_map_44 (duu_0,dvv_0)) (zipt (u_0,du_1,wet_s_1,v_0,dv_1))
+         -- dyn_map_64
+        s5 = [-1,0]
+        vn_s_0 = stencil s5 vn_1
+        s6 = [-502,0]
+        un_s_0 = stencil s6 un_1
+        s7 = [-1,-502,0,502,1]
+        h_s_0 = stencil s7 h_0
+        etan_1 =  map (dyn_map_64 (dt_0,dx_0,dy_0)) (zipt (un_s_0,h_s_0,vn_s_0,eta_s_0))
+         -- vernieuw_map_23
+        (h_1,u_1,v_1,wet_1) = unzipt $ map (vernieuw_map_23 hmin_0) (zipt (hzero_0,eta_s_0,h_s_0,un_s_0,vn_s_0))
+      in
+        (eta_1,du_1,dv_1,un_1,vn_1,etan_1,h_1,u_1,v_1,wet_1)
+-}
+
+type Name = String
+data VE = VI  | VO  | VS  | VT deriving (Show, Typeable, Data, Eq)
+type AST = [(Expr,Expr)]                      
+
+data Expr =
+        -- Left-hand side:
+                      Scalar 
+                    | Tuple [Expr]
+                    | Vec VE Name
+
+        -- Right-hand side:
+                    | SVec Int Name
+                    | ZipT [Expr] 
+                    | UnzipT Expr
+                    | Elt Int Expr
+                    | PElt Int
+                    | Map Expr Expr
+                    | Fold Expr Expr Expr
+                    | Stencil Expr Expr 
+                    | Function Name 
[…]
