+++ TyTraCLTransformations/memory-reduction-TyTraCL.txt
@@ -1,0 +1,224 @@
+2018-07-27
+
+I realised that smart caching can be useful on GPUs even though we don't stream the calculations. 
+
+The key point is that in Tytra, we assume that the input data is in memory and that for iterative operations we use a double buffer. 
+
+Currently, the LES uses a large number of intermediate arrays which are stored in the global memory. 
+
+I contend that we only need the actual input arrays and we can recompute all the rest from those. So rather than writing back to memory, we store what we need in a smart buffer.
+
+For the LES we have u,v,w,p as inputs, but we actually use:
+
+p (double)
+u,v,w
+usum,vsum,wsum
+f,g,h
+fold,gold,hold
+sm
+rhs
+
+
+So my contention is that we could go from these 16 arrays to only 4 plus 12 smart buffers which only take 3/kp of the space, so in total:
+
+4+12*3/kp = 4.45 for kp=80 so a reduction in memory of 3.6x, so we could go from the current domain size of 300x300 to 568x568
+
+
+However, this requires staging. For example, take the case of the rhs for the SOR. 
+
+This is calculated from p,u,v,w using a reduction, and then the reduced value is subtracted from the original. 
+
+So what we have to do is always recompute rather than store the intermediates in memory. 
+
+Now, let's look at the following:
+
+map : (a -> b) -> Vect n a -> Vect n b
+mapt : (a -> b) -> Tup k a -> Tup k b
+stencil : Tup k Int -> Vect n a -> Vect n (Tup k a)
+
+
+
+s1 : Tup k Int
+f1 : a -> b
+
+stencil s1 (map f1) = map (mapt f1) (stencil s1)
+
+v0 : Vect n a
+f1 : a -> b
+v1 : Vect n b
+
+v1s : Vect n (Tup k b)
+v1 = map f1 v0
+v1s = stencil s1 v1
+
+v0s : Vect (Tup k a)
+v0s = stencil s1 v0
+
+mapt f1 : Tup k a -> Tup k b
+v1s : Vect n (Tup k b)
+v1s = map (mapt f1) v0s 
+
+or
+
+stencil s1 (map f1 v1) = map (mapt f1) (stencil s1 v1)
+
+
+A few other transforms and names
+
+
+applyt (f1,f2) (e1,e2) =\(e1,e2) -> (f1 e1,f2 e2)
+
+zipt (map f1 v1, map f2 v2) = map (applyt (f1,f2)) (zipt (v1, v2))
+
+
+applyt (g1,g2) $ applyt (f1,f2)  = applyt (g1 . f1, g2 . f2)
+
+----
+
+
+How do we implement `stencil` in GPU OpenCL?
+
+    map (mapt f) (stencil s v)
+
+could maybe be written using a stencilmap primitive:
+
+    stencilmap s f v = map (mapt f) (stencil s v)
+
+
+
+The implementation in OpenCL would be for example like this:
+
+for offset_idx in stencil:
+    f(v[idx+offset_idx) 
[…]

[…]
