### 2018-11-23 81d7d0c
Rewrite rules for memory reduction on TyTraCL. Most of the code is written, but the rules need to be applied until a fixed point is reached.
