+++ TyTraCLTransformations/memory-reduction-TyTraCL.hs
@@ -11,14 +11,21 @@ main = do
     mapM print ast'
     putStrLn "\nSubstitute vectors (recursive)"
     mapM print ast''
+    putStrLn "\nApply rewrite rules"
+    mapM print ast'''
+{-    
     putStrLn "\nTest for Vec in RHS Expr"
     mapM (print . get_vec_subexprs . snd) ast''
     putStrLn "\nTest for non-input Vec in RHS Expr"
     mapM (print . expr_has_non_input_vecs . snd) ast'
     putStrLn "\nTest for non-input Vec in RHS Expr"
     mapM (print . expr_has_non_input_vecs . snd) ast''
+    -- Tests for bottom-up reduction using everywhere
+    print ll
+    print (reduce_ll ll)
+    print (reduce_tree tree)
 --    print $ find_in_ast ast' (Vec VT "duu_1")
-
+-}
 {-
     main (etan_0,wet_0,eps_0,dt_0,g_0,dx_0,dy_0,wet_0,u_0,v_0,duu_0,dvv_0,h_0,dt_0,dx_0,dy_0,hzero_0,hmin_0) =
       let
@@ -56,7 +63,8 @@ type AST = [(Expr,Expr)]
 
 data Expr =
         -- Left-hand side:
-                      Scalar 
+                      Scalar Name
+                    | Const Int  
                     | Tuple [Expr]
                     | Vec VE Name
 
@@ -70,8 +78,9 @@ data Expr =
                     | Fold Expr Expr Expr
                     | Stencil Expr Expr 
                     | Function Name 
-                    | ApplyT [Name]  
-                    | MapS Name
+                    | ApplyT [Expr]  
+                    | MapS Expr
+                    | Comp Expr Expr
                         deriving (Show, Typeable, Data, Eq)
 
         -- Initially, all RHS nodes will be Map, Fold or Stencil.
@@ -180,9 +189,20 @@ Finally, we go trough all lines of the ast and do a recursive substitution:
 -}    
 ast'' =  map (substitute_vec_rec ast') (filter lhs_is_output_vec ast')
 
+{-
+The AST is now reduced to a list of tuples where the first elt (LHS) is an output vector and the second element is an expression which only contains input vectors.     
+Now we should start applying the rewrite rules to reduce each of these expressions to a single Map.
+-}
+
+
+
+
+
+
 lhs_is_output_vec (lhs_vec,expr) = case lhs_vec of
     Vec VO _ -> True
     _ -> False
+
 -- Create a list of all Vecs in the expr
 -- Find the corresponding RHS expressions
 -- Substitute all of them in the given expression
@@ -206,3 +226,92 @@ substitute_vec_rec ast expr_tup@(lhs_vec,expr) = let
                 substitute_vec_rec ast expr_tup'
         else -- we're dome, return the result 
             expr_tup
+
+-- A linked list
+ll = Tuple [ Const 11, Tuple [ Const 22, Tuple [ Const 33, Tuple [ Const 44, Tuple [   Const 55, Const 66 ] ] ] ] ]
+
+
+-- Sum all nodes in the linked list
+reduce_ll ll = everywhere (mkT reduce_ll_chain) ll
+
+reduce_ll_chain expr = case expr of
+    Tuple [ Const n, Const m] -> Const (n+m)
+    _ -> expr
+
+
+{-
+ A binary tree
+                    11
+                /       \
+             22            44    
+           /   \        /     \
+          33    0      55      66
[…]
