### 2018-12-08 a665ad1
Progress on rewriting the program for memory reduction. Main remaining task is the Fortran emitter.
