+++ TyTraCLTransformations/memory-reduction-TyTraCL.hs
@@ -1,8 +1,10 @@
 {-# LANGUAGE DeriveDataTypeable #-}
 module Main where
 
+import Control.Monad.State
+
 --import Data.Generics (mkQ, mkT, gmapQ, gmapT, everything, everywhere)
-import Data.Generics (Data, Typeable, mkQ, mkT, gmapQ, gmapT, everything, everywhere, everywhere')
+import Data.Generics (Data, Typeable, mkQ, mkT, mkM, gmapQ, gmapT, everything, everywhere, everywhere', everywhereM)
 
 main = do
     putStrLn "Original AST"
@@ -13,6 +15,9 @@ main = do
     mapM print ast''
     putStrLn "\nApply rewrite rules"
     mapM print ast'''
+--    mapM print map_checks
+    putStrLn "\nDecompose expressions"
+    mapM (mapM print) ast4
 {-    
     putStrLn "\nTest for Vec in RHS Expr"
     mapM (print . get_vec_subexprs . snd) ast''
@@ -59,28 +64,34 @@ main = do
 
 type Name = String
 data VE = VI  | VO  | VS  | VT deriving (Show, Typeable, Data, Eq)
+    
 type AST = [(Expr,Expr)]                      
 
 data Expr =
         -- Left-hand side:
                       Scalar Name
-                    | Const Int  
-                    | Tuple [Expr]
-                    | Vec VE Name
+                    | Const Int -- bb: IntLit Integer
+                    | Tuple [Expr] --  bb: Tup [Expr]
+                    | Vec VE Name -- bb: Var Name, type via cofree comonad, but VE info is not there
 
         -- Right-hand side:
-                    | SVec Int Name
-                    | ZipT [Expr] 
-                    | UnzipT Expr
-                    | Elt Int Expr
-                    | PElt Int
-                    | Map Expr Expr
-                    | Fold Expr Expr Expr
-                    | Stencil Expr Expr 
-                    | Function Name 
-                    | ApplyT [Expr]  
-                    | MapS Expr
-                    | Comp Expr Expr
+                    | SVec Int Name -- bb: SVec [Expr] -> to get a name, use a Let
+                    | ZipT [Expr] -- bb: App Zip (Tup  [...])
+                    | UnzipT Expr -- bb: App Unzip (vec of tuples)
+                    | Elt Int Expr -- bb: App (Select Integer) Tup
+                    | PElt Int -- bb does not need this
+                    | Map Expr Expr -- bb: App (Map Expr) Expr
+                    | Fold Expr Expr Expr -- bb: App (Fold (App action acc) Expr
+                    | Stencil Expr Expr -- bb uses App : App (Stencil (SVec [IntLit])) vector
+                    | Function Name -- bb: uses Var Name with a function type
+                    | Id -- bb has Id 
+                    | Mu Expr Expr -- \a e -> g a (f e) -- of course bb does not have this, no need
+                    | ApplyT [Expr]  -- bb: App FTup [Expr]
+                    | MapS Expr -- bb does not have this, not needed
+                    | Comp Expr Expr -- bb does not have this, not needed
+                    -- | bb has Let Expr Expr
+                    -- | bb has App Expr Expr to apply and Expr to an Expr
[…]
