### 2018-12-12 #865 @wimvanderbauwhede
Thanks a lot! I've released the latest tagged version (V1.1.1) and archived it on Zenodo. The DOI is

10.5281/zenodo.2222337
