### 2018-12-12 #865 @labarba
You will need to manually fix the author list in the Zenodo archive.
