### 2018-12-12 #865 @labarba
In addition to the authors, the title might need manual editing, too. Zenodo just grabs its metadata from the repo.
