### 2018-12-12 #865 @wimvanderbauwhede
Thanks, I didn't realise that. Title and author is now as in the JOSS paper.
