### 2018-12-12 #865 @wimvanderbauwhede
OK, I did 0., 1. and 2.
DOI resolves, have downloaded and checked the PDF. On to 3. !
Thank you very much for your patience and assistance!
